Add files via upload
Race Director helps when making the video file. It'll tell you which events are coming up in the next 20 seconds so that you can switch the view to that car. But, only use this after the filtering process as there will be too many events and you will be overwhelmed. 

The spline data shower is a debugging tool which I've been using to identify spikes in the spline data, but could be developed into something to visualize a race too.